Conflicts:
	app/assets/stylesheets/application.css.scss
	app/controllers/events/registrations_controller.rb
	app/models/registration.rb
	app/views/layouts/dashboard.html.erb
	db/schema.rb

+++ db/migrate/20140223080947_create_events.rb
@@ -5,7 +5,7 @@ class CreateEvents < ActiveRecord::Migration
 			t.integer :event_type
 			t.date :begin_date
 			t.date :end_date
-			t.decimal	:registration_cost
+			t.integer	:registration_cost
 		  t.date :registration_open_date
 			t.date :registration_close_date
 			t.references :location, index: true
